Restructured project code
