--- backend/src/main.js
@@ -1,6 +1,0 @@
-import { initializeMap } from '../map/mapSetup.js';
-
-document.addEventListener('DOMContentLoaded', () => {
-    initializeMap();
-});
-
