--- frontend/src/mapClickHandlers.js
@@ -1,41 +1,0 @@
-import { measurePoints } from "../../backend/threejs/measurePoints.js";
-
-mapboxgl.accessToken = 'pk.eyJ1IjoibXljc2FsIiwiYSI6ImNsc2RtM2tvdzEyNnIybXQwcjI5d2tqcjAifQ.SqGe3A-JLNSkTCYluSpRnA';
-
-export let lng;
-export let lat;
-
-const lngDisplay = document.getElementById('lng');
-const latDisplay = document.getElementById('lat');
-const eleDisplay = document.getElementById('ele');
-
-export const handleMapClick = (event, map, tb) => {
-    // When the map is clicked, set the lng and lat
-    // equal to the lng and lat properties in the returned lngLat object.
-    lng = event.lngLat.lng;
-    lat = event.lngLat.lat;
-    getElevation();
-    measurePoints(event, map, tb);
-
-};
-
-async function getElevation() {
-    // Construct the API request.
-    const query = await fetch(
-        `https://api.mapbox.com/v4/mapbox.mapbox-terrain-v2/tilequery/${lng},${lat}.json?layers=contour&limit=50&access_token=${mapboxgl.accessToken}`,
-        { method: 'GET' }
-    );
-    if (query.status !== 200) return;
-    const data = await query.json();
-    // Display the longitude and latitude values.
-    lngDisplay.textContent = lng.toFixed(8);
-    latDisplay.textContent = lat.toFixed(8);
-    // Get all the returned features.
-    const allFeatures = data.features;
-    // For each returned feature, add elevation data to the elevations array.
-    const elevations = allFeatures.map((feature) => feature.properties.ele);
-    // In the elevations array, find the largest value.
-    const highestElevation = Math.max(...elevations);
-    // Display the largest elevation value.
-    eleDisplay.textContent = `${highestElevation} meters`;
-}
